pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-gtd-benchmark"
version = "0.1.0"
description = "A rigorous benchmark suite for LLM-based tabular data generation models."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }

dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "scipy>=1.11",
    "scikit-learn>=1.3",
    "xgboost>=1.7",
]

[project.optional-dependencies]
# Optional: much faster KNN on large datasets (Linux/macOS preferred)
faiss = ["faiss-cpu>=1.7"]
dev = ["pytest>=7", "pytest-cov", "ruff"]

[tool.setuptools.packages.find]
where = ["."]
include = ["llm_gtd_benchmark*"]

[tool.ruff]
line-length = 100
target-version = "py39"
